+++ src/components/Main.tsx
@@ -1,13 +1,10 @@
 // External Libraries
-import React from 'react';
 import { useState } from 'react';
 import { useZxing } from 'react-zxing';
 
 // googlespreadsheet
 import { reflectLocal } from './localToGSsheet';
 import { UserComponent } from './user';
-// import { LoginErrorComponent } from './loginError';
-import { deleteRowFromSheet } from './SheetOperater';
 
 // localStrage
 // import { local_key_array } from './localStorageLib';
@@ -18,17 +15,12 @@ const localStorageLib = new LocalStorageLib();
 // import './style.css';
 import CSVTableComponent2, { CSVDownloadButton1 } from './csvDownload';
 import { TotalTable } from './totalTable';
+import { DataTable } from './DataTable';
 import { productData } from './data';
 import { CreateCal } from './showCal';
 
 // Material UI
-import { BottomNavigation, BottomNavigationAction, Box, IconButton } from '@mui/material';
-import Table from '@mui/material/Table';
-import TableBody from '@mui/material/TableBody';
-import TableCell from '@mui/material/TableCell';
-import TableContainer from '@mui/material/TableContainer';
-import TableHead from '@mui/material/TableHead';
-import TableRow from '@mui/material/TableRow';
+import { BottomNavigation, BottomNavigationAction, Box} from '@mui/material';
 import Paper from '@mui/material/Paper';
 
 // Material Icons
@@ -36,7 +28,6 @@ import CalculateIcon from '@mui/icons-material/Calculate';
 import DataThresholdingIcon from '@mui/icons-material/DataThresholding';
 import QrCodeIcon from '@mui/icons-material/QrCode2';
 import AccessibilityNewIcon from '@mui/icons-material/AccessibilityNew';
-import { Delete as DeleteIcon } from '@mui/icons-material';
 
 // Default Data
 let products = [
@@ -54,15 +45,6 @@ interface SellItem {
 }
 
 function App() {
-  // const _productData: Item2[] = productData.map((data) => {
-  //   const one_of_productData = {
-  //     ...data,
-  //     price: Number(data.price),
-  //     quantity: 0
-  //   }
-  //   return one_of_productData;
-  // });
-  // const [all_products, setAll_products] = useState<Item2[]>(_productData);
 
   let timeArray: string[] = ["9:00", "10:00", "11:00"]; // 時間の配列
   let quantityArray: string[] = ["3", "5", "2"]; // 品物の個数の配列
@@ -78,58 +60,6 @@ function App() {
   });;
   const [_SellItem, setSellIetm] = useState<SellItem[]>(set_array);
 
-  const [data, setData] = useState<Item[]>(items);
-  const DataTable: React.FC<{ items: Item[] }> = () => {
-
-
-    const handleDelete = (index: number, time: string) => {
-      const confirmed = window.confirm(`データ「${time}」を削除しますか？`);
-      if (!confirmed) return;
-      const newData = [...data];
-      newData.splice(index, 1);
-      setData(newData);
-
-      const getKey = Object.keys(localStorage);
-      for (let i = 0; i < getKey.length; i++) {
-        if (getKey[i].indexOf(time) >= 0) {
-          localStorage.removeItem(getKey[i]);
-        }
-      }
-
-      deleteRowFromSheet(time);
-    };
-
-    return (
-      <TableContainer sx={{ maxWidth: 500 }} component={Paper}>
-        <Table size='small' aria-label='simple table'>
-          <TableHead>
-            <TableRow>
-              <TableCell>時間</TableCell>
-              <TableCell>品物の個数</TableCell>
-              <TableCell>共有</TableCell>
-              <TableCell></TableCell>
-            </TableRow>
-          </TableHead>
-          <TableBody>
-            {data.map((item, index) => (
-              <TableRow key={index}>
-                <TableCell>{item.time}</TableCell>
-                <TableCell>{item.quantity}</TableCell>
-                <TableCell></TableCell>
-                <TableCell>
-                  <IconButton
-                    color="primary"
-                    onClick={() => { handleDelete(index, item.time); updateData(); reflectLocal(); }}>
-                    <DeleteIcon />
-                  </IconButton>
-                </TableCell>
-              </TableRow>
-            ))}
-          </TableBody>
-        </Table>
-      </TableContainer>
-    );
-  };
 
   const updateData = () => {
     timeArray = [];
@@ -151,7 +81,6 @@ function App() {
       }
     }
     items = timeArray.map((time, index) => ({ time, quantity: quantityArray[index] }));
-    setData(timeArray.map((time, index) => ({ time, quantity: quantityArray[index] })));
 
     setSellIetm(set_array);
     const sophisticatedQuantityArray: string[][] = [];
@@ -203,32 +132,7 @@ function App() {
         _qtyArray[i] = qty;
         _sumArray[i] = eachSum;
       }
-
-      // local strage
-      // const keys  = Object.keys(localStorage);
-      // let d = new Date();
-      // let year = d.getFullYear();
-      // let month = d.getMonth() + 1;
-      // let day = d.getDate();
-      // let hour = d.getHours().toString().padStart(2, '0');
-      // let minute = d.getMinutes().toString().padStart(2, '0');
-      // let seconds = d.getSeconds().toString().padStart(2, '0');
-      // let UTCtime = d.getTime().toString().slice(0, -3);
-      // const date = UTCtime + ")" + year + '-' + month + '-' + day + ' ' + hour + ':' + minute + ':' + seconds;
-      //console.log(date);
-
-      // all_products.splice(0);
-      // for (let i = 0; i < _nameArray.length - 1; i++) {
-      //   all_products.push(
-      //     {
-      //       product: _nameArray[i],
-      //       price: _costArray[i],
-      //       quantity: _qtyArray[i]
-      //     }
-      //   );
-      // }
-      //console.log(all_products);//showCalに渡したい
-      // setAll_products(all_products);
+      
       //表示する商品
       products = [];
       for (let i = 0; i < _nameArray.length - 1; i++) {
@@ -360,7 +264,8 @@ function App() {
           <TotalTable />
           <hr />
           {/* <CSVTableComponent2 data={data} /> */}
-          <DataTable items={data} />
+          {/* <DataTable items={data} /> */}
+          <DataTable/>
         </div>
       }
